saturday updates

+++ 2016Robot/src/org/usfirst/frc/team5915/robot/commands/MoveArm.java
@@ -8,9 +8,9 @@ import edu.wpi.first.wpilibj.command.Command;
  *
  */
 public class MoveArm extends Command {
-	private int dir = 0;
+	private double dir = 0;
 
-    public MoveArm(int direction) {
+    public MoveArm(double direction) {
     	requires(Robot.manipulatorArm);
 		dir = direction;
     }
